App: Fixed crash caused by deletion of local images after upload. They got a new path

+++ app/src/main/java/com/buur/frederik/multimediechatexample/fragments/chatfragment/ChatController.kt
@@ -165,7 +165,9 @@ class ChatController {
                                             ?.observeOn(AndroidSchedulers.mainThread())
                                 }
                                 .doOnNext { uploadResponse ->
-                                    MMData.deleteFile(mmData.source)
+                                    if (mmData.type == MMDataType.Audio.ordinal) {
+                                        MMData.deleteFile(mmData.source)
+                                    }
                                     mmData.source = uploadResponse.url
                                     val gson = Gson().toJson(mmData)
                                     socket?.emit(TOPIC_NEW_MESSAGE, gson)
